add batch books delete signature
