+++ internal/server/users.go
@@ -1,0 +1,110 @@
+package server
+
+import (
+	"errors"
+	"net/http"
+
+	"github.com/Dorrrke/g3-bookly/internal/domain/models"
+	"github.com/Dorrrke/g3-bookly/internal/logger"
+	storerrros "github.com/Dorrrke/g3-bookly/internal/storage/errros"
+	"github.com/gin-gonic/gin"
+)
+
+func (s *Server) register(ctx *gin.Context) {
+	log := logger.Get()
+	var user models.User
+	if err := ctx.ShouldBindBodyWithJSON(&user); err != nil {
+		log.Error().Err(err).Msg("unmarshal body failed")
+		ctx.String(http.StatusBadRequest, "incorrectly entered data")
+		return
+	}
+	if err := s.valid.Struct(user); err != nil {
+		log.Error().Err(err).Msg("validate user failed")
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	uuid, err := s.storage.SaveUser(user)
+	if err != nil {
+		if errors.Is(err, storerrros.ErrUserExists) {
+			log.Error().Msg(err.Error())
+			ctx.String(http.StatusConflict, err.Error())
+			return
+		}
+		log.Error().Err(err).Msg("save user failed")
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	log.Debug().Str("uuid", uuid).Send()
+	token, err := s.createJWTToken(uuid)
+	if err != nil {
+		log.Error().Err(err).Msg("create jwt failed")
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	ctx.Header("Authorization", token)
+	ctx.String(http.StatusCreated, uuid)
+}
+
+func (s *Server) login(ctx *gin.Context) {
+	log := logger.Get()
+	var user models.User
+	if err := ctx.ShouldBindBodyWithJSON(&user); err != nil {
+		log.Error().Err(err).Msg("unmarshal body failed")
+		ctx.String(http.StatusBadRequest, "incorrectly entered data")
+		return
+	}
+	if err := s.valid.Struct(user); err != nil {
+		log.Error().Err(err).Msg("validate user failed")
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+	uuid, err := s.storage.ValidUser(user)
+	if err != nil {
+		if errors.Is(err, storerrros.ErrUserNoExist) {
+			log.Error().Err(err).Msg("user not found")
+			ctx.String(http.StatusNotFound, "invalid login or password: %w", err)
+			return
+		}
+		if errors.Is(err, storerrros.ErrInvalidPassword) {
+			log.Error().Err(err).Msg("invalid pass")
+			ctx.String(http.StatusUnauthorized, err.Error())
+			return
+		}
+		log.Error().Err(err).Msg("validate user failed")
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	token, err := s.createJWTToken(uuid)
+	if err != nil {
+		log.Error().Err(err).Msg("create jwt failed")
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+	ctx.Header("Authorization", token)
+	ctx.String(200, "user %s are logined", uuid)
+}
+
+func (s *Server) userInfo(ctx *gin.Context) {
+	log := logger.Get()
+	toketn := ctx.GetHeader("Authorization")
+	if toketn == "" {
+		ctx.String(http.StatusUnauthorized, "invalid token")
+		return
+	}
+	uid, err := validToken(toketn)
+	if err != nil {
+		log.Error().Err(err).Msg("validate jwt failed")
+		ctx.String(http.StatusUnauthorized, "invalid token")
+		return
+	}
+	user, err := s.storage.GetUser(uid)
+	if err != nil {
+		if errors.Is(err, storerrros.ErrUserNotFound) {
+			ctx.String(http.StatusNotFound, err.Error())
+			return
+		}
+		ctx.String(http.StatusInternalServerError, err.Error())
+		return
+	}
+	ctx.JSON(http.StatusFound, user)
+}
